requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
numpy==1.17.0
plotly==4.1.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
retrying==1.3.3
rsa==4.0
scipy==1.3.0
six==1.12.0
tqdm==4.33.0
urllib3==1.25.3
